Refactor: make separate components for modularity - make separate component for connect wallet button - make separate component for connect arcana button - move uploadFile under arcana folder
